Fixed bugs in setting column plot designation.
git-svn-id: http://svn.berlios.de/svnroot/repos/qtiplot/trunk@2011 5a6a7de5-fb12-0410-b871-c33778c25c60

+++ qtiplot/src/core/ApplicationWindow.cpp
@@ -7722,7 +7722,7 @@ void ApplicationWindow::showColMenu(int c)
 	QMenu stat(this);
 	QMenu norm(this);
 
-	if ((int)w->selectedColumns().count() == 1){
+	if (w->selectedColumns().count() == 1){
 		w->setSelectedCol(c);
 		plot.addAction(actionPlotL);
 		plot.addAction(actionPlotP);
@@ -7853,7 +7853,7 @@ void ApplicationWindow::showColMenu(int c)
 		contextMenu.insertSeparator();
 		contextMenu.addAction(actionAdjustColumnWidth);
 		contextMenu.addAction(actionShowColumnOptionsDialog);
-	} else if ((int)w->selectedColumns().count() > 1){
+	} else if (w->selectedColumns().count() > 1){
 		plot.addAction(actionPlotL);
 		plot.addAction(actionPlotP);
 		plot.addAction(actionPlotLP);
